+++ services/soso-web/src/types/interfaces.ts
@@ -1,0 +1,7 @@
+export interface User {
+  id: string;
+  username: string;
+  mailAddress: string;
+  hasCar: boolean;
+  capacity: number;
+}
