+++ services/soso-web/src/views/LoginView/components/GoToRegisterLink.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
 
-export const RegisterButton: React.FC = () => {
+export const GoToRegisterLink: React.FC = () => {
   return (
     <div className="text-center mt-4">
       <p className="text-sm text-slate-600 mb-2">アカウントをお持ちでない方</p>
